Update README.md
A simple web application to manage blog posts, built with Node.js and Express.js. This application allows users to create, edit, and delete blog posts. The project demonstrates the use of server-side rendering with EJS, as well as handling HTTP requests for CRUD operations.

Features=>
Home Page: Displays a list of all blog posts with options to edit or delete each post.
Create New Post: Allows users to add a new blog post with a title and content.
Edit Post: Enables users to modify the title and content of an existing post.
Delete Post: Provides the functionality to remove a specific post.